Replace the city text field by a drop-down of cities for a specific country in Woo-commerce
The following code (jQuery powered) will replace the city text field by a custom dropdown of cities for a specific country only and for this specific country, if the city selected value is "Others", an additional text field will appear under the cities dropdown, where customer can enter manually a different city.

The code works for shipping and billing fields independently.

When "Others" is selected for the defined country, the two last functions will:

Validate that the city additional field is filled,
Save the city value as billing or shipping city value.

Code goes in function.php file of your active child theme (active theme). Tested and works.

Please Remember that a customer can be in a foreign country outside Saudi Arabia (billing country) and buy something that will be shipped in Saudi Arabia (shipping country).