Create how much to go for free shipping
Everybody wants to increase the average order value for their store, right? A good way to do this is to offer free shipping above a certain order amount. If you have this on your site you should of course promote this as a selling point on a prominent and default position.

Alongside with that, it would be good to help the customer remind in the cart if they’re close to the free shipping amount with a notification saying something like “Add $10 more to your basket to receive free shipping!”.

Showing a free shipping notice for WooCommerce Core free shipping rates
Using the following snippet you can show the notice on the cart / checkout page. It will only appear when the minimum order amount setting is set.